Fri 25 Apr 2025: 1st commit

+++ LMS-Mangment-Service/src/main/java/com/slt/peotv/lmsmangmentservice/repository/AccessLogArchiveRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccessLogArchiveRepo extends JpaRepository<AccessLogArchiveEntity, Long> {
+public interface AccessLogArchiveRepo extends JpaRepository<AccessLogArchiveEntity, Integer> {
 }
